split component for options creation from data
